requirements.txt
tensorflow-macos==2.10.0
tensorflow-metal==0.6.0
matplotlib==3.6.2
numpy==1.23.1
gym==0.23.1